+++ src/idt/irq.c
@@ -1,5 +1,6 @@
 #include <kernel/io.h>
 #include <kernel/ints.h>
+#include <stdint.h>
 
 extern void irq0();
 extern void irq1();
@@ -35,22 +36,22 @@ static void irq_remap(void)
 void irq_install()
 {
 	irq_remap();
-	idt_set_gate(32,(unsigned)irq0,0x08,0x8E);
-	idt_set_gate(33,(unsigned)irq1,0x08,0x8E);
-	idt_set_gate(34,(unsigned)irq2,0x08,0x8E);
-	idt_set_gate(35,(unsigned)irq3,0x08,0x8E);
-	idt_set_gate(36,(unsigned)irq4,0x08,0x8E);
-	idt_set_gate(37,(unsigned)irq5,0x08,0x8E);
-	idt_set_gate(38,(unsigned)irq6,0x08,0x8E);
-	idt_set_gate(39,(unsigned)irq7,0x08,0x8E);
-	idt_set_gate(40,(unsigned)irq8,0x08,0x8E);
-	idt_set_gate(41,(unsigned)irq9,0x08,0x8E);
-	idt_set_gate(42,(unsigned)irq10,0x08,0x8E);
-	idt_set_gate(43,(unsigned)irq11,0x08,0x8E);
-	idt_set_gate(44,(unsigned)irq12,0x08,0x8E);
-	idt_set_gate(45,(unsigned)irq13,0x08,0x8E);
-	idt_set_gate(46,(unsigned)irq14,0x08,0x8E);
-	idt_set_gate(47,(unsigned)irq15,0x08,0x8E);
+	idt_set_gate(32,(uint64_t)irq0,0x08,0x8E);
+	idt_set_gate(33,(uint64_t)irq1,0x08,0x8E);
+	idt_set_gate(34,(uint64_t)irq2,0x08,0x8E);
+	idt_set_gate(35,(uint64_t)irq3,0x08,0x8E);
+	idt_set_gate(36,(uint64_t)irq4,0x08,0x8E);
+	idt_set_gate(37,(uint64_t)irq5,0x08,0x8E);
+	idt_set_gate(38,(uint64_t)irq6,0x08,0x8E);
+	idt_set_gate(39,(uint64_t)irq7,0x08,0x8E);
+	idt_set_gate(40,(uint64_t)irq8,0x08,0x8E);
+	idt_set_gate(41,(uint64_t)irq9,0x08,0x8E);
+	idt_set_gate(42,(uint64_t)irq10,0x08,0x8E);
+	idt_set_gate(43,(uint64_t)irq11,0x08,0x8E);
+	idt_set_gate(44,(uint64_t)irq12,0x08,0x8E);
+	idt_set_gate(45,(uint64_t)irq13,0x08,0x8E);
+	idt_set_gate(46,(uint64_t)irq14,0x08,0x8E);
+	idt_set_gate(47,(uint64_t)irq15,0x08,0x8E);
 }
 
 void *irq_routines[16] =
